+++ engine/apps/zvonok/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import CallStatusCallback
+
+urlpatterns = [
+    path("call_status_events", CallStatusCallback.as_view(), name="call_status_events"),
+]